+++ proxy/src/usage_metrics.rs
